+++ app/components/exam/ExamForm.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { actionCreators } from '../../actions/index'
 import { sortData } from '../rooms/ClassList'
-import examForm from './ExamFormHelper'
+import examForm from './helpers/formHelper'
 
 const _ = require('lodash')
