+++ app/components/exam/helpers/formHelper.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles/exam.css'
+import styles from '../../styles/exam.css'
 
 const examForm = (subjectOptions, classOption, actions) => {
 	const examFields = (
