+++ app/components/students/helpers/modalHelper.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles/students.css'
+import styles from '../../styles/students.css'
 
 const _ = require('lodash')
